+++ examples/histogram-torus/run.py
@@ -31,7 +31,7 @@ name = args.name
 
 # Parse the compile metadata
 compile_data = None
-with open(f"{name}/out.json") as json_file:
+with open(f"{name}/out.json", encoding="utf-8") as json_file:
   compile_data = json.load(json_file)
 assert compile_data is not None
 compile_colors = compile_data["colors"]
@@ -47,7 +47,6 @@ OUT_COLOR = int(compile_colors["OUT_COLOR"])
 np.random.seed(seed=7)
 
 elfs = glob(f"{name}/bin/out_[0-9]*.elf")
-sim_out_path = f"{name}/bin/core.out"
 runner = CSELFRunner(elfs, cmaddr=args.cmaddr)
 
 # Randomly generate the input values.
@@ -80,7 +79,7 @@ for i in range(HIST_W):
 out_port_map = f"{{out_tensor[idx=0:0] -> [PE[{HIST_W-1},-1] -> index[idx]]}}"
 runner.add_output_tensor(OUT_COLOR, out_port_map, np.uint32)
 
-runner.connect_and_run(sim_out_path)
+runner.connect_and_run()
 
 # check that tally output was as expected
 result_tensor = runner.out_tensor_dict["out_tensor"]