+++ tutorials/17-libraries/commands.sh
@@ -1,2 +1,7 @@
-cslc ./code.csl --fabric-dims=3,3 --fabric-offsets=1,1 -o out
-cs_python run.py --name out
+#!/usr/bin/env bash
+
+set -e
+
+cslc ./code.csl --fabric-dims=3,3 --fabric-offsets=1,1 \
+--params=iterations:200 -o out
+cs_python run.py --name out --tolerance 0.1