+++ tutorials/17-libraries/run.py
@@ -23,25 +23,25 @@ from cerebras.elf.cs_elf_runner import CSELFRunner
 parser = argparse.ArgumentParser()
 parser.add_argument('--name', help='the test name')
 parser.add_argument("--cmaddr", help="IP:port for CS system")
+parser.add_argument("--tolerance", type=float, help="tolerance for result")
 args = parser.parse_args()
 name = args.name
 
-# Path to ELF and simulation output files
+# Path to ELF files
 elf_paths = [f"{name}/bin/out_0_0.elf"]
-sim_out_path = f"{name}/bin/core.out"
 
-# Simulate ELF file and produce the simulation output
+# Simulate ELF files
 runner = CSELFRunner(elf_paths, cmaddr=args.cmaddr)
 
 result_color = 1
-result_port_map = f"{{result_out[idx=0:0] -> [PE[1,0] -> index[idx]]}}"
+result_port_map = "{result_out[idx=0:0] -> [PE[1,0] -> index[idx]]}"
 runner.add_output_tensor(result_color, result_port_map, np.float32)
 
 tsc_color = 2
-tsc_port_map = f"{{tsc_out[idx=0:5] -> [PE[1,0] -> index[idx]]}}"
+tsc_port_map = "{tsc_out[idx=0:5] -> [PE[1,0] -> index[idx]]}"
 runner.add_output_tensor(tsc_color, tsc_port_map, np.uint16)
 
-runner.connect_and_run(sim_out_path)
+runner.connect_and_run()
 result = runner.out_tensor_dict["result_out"]
 
 # Helper functions for computing the delta in the cycle count
@@ -54,5 +54,5 @@ def subtract_timestamps(words):
 cycles = subtract_timestamps(runner.out_tensor_dict["tsc_out"])
 print("cycle count:", cycles)
 
-np.testing.assert_allclose(result, np.pi, atol=0.01, rtol=0)
+np.testing.assert_allclose(result, np.pi, atol=args.tolerance, rtol=0)
 print("SUCCESS!")