+++ tutorials/19-fifos/run.py
@@ -28,18 +28,17 @@ parser.add_argument("--cmaddr", help="IP:port for CS system")
 args = parser.parse_args()
 name = args.name
 
-# Path to ELF and simulation output files
+# Path to ELF files
 elf_paths = glob(f"{name}/bin/out_*.elf")
-sim_out_path = f"{name}/bin/core.out"
 
-# Simulate ELF file and produce the simulation output
+# Simulate ELF files
 runner = CSELFRunner(elf_paths, cmaddr=args.cmaddr)
 
 np.random.seed(seed=7)
 
 # Parse the compile metadata
 compile_data = None
-with open(f"{name}/out.json") as json_file:
+with open(f"{name}/out.json", encoding="utf-8") as json_file:
   compile_data = json.load(json_file)
 assert compile_data is not None
 compile_params = compile_data["params"]
@@ -53,7 +52,7 @@ output_port_map = f"{{out_tensor[idx=0:{size - 1}] -> [PE[1,0] -> index[idx]]}}"
 output_color = 3 # The output color to which the kernel will send data.
 runner.add_output_tensor(output_color, output_port_map, np.float16)
 
-runner.connect_and_run(sim_out_path)
+runner.connect_and_run()
 
 result_tensor = runner.out_tensor_dict["out_tensor"]